consultancy/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
stripe==11.4.1
reportlab==4.2.5
python-multipart==0.0.20
python-dotenv==1.0.1
Jinja2==3.1.4
pydantic==2.10.4
